% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_fwd.R
\name{add_edge}
\alias{add_edge}
\title{Add edge to an adjacency matrix}
\usage{
add_edge(G, v1, v2)
}
\arguments{
\item{G}{Adjacency matrix.}

\item{v1}{Vertex 1.}

\item{v2}{Vertex 2.}
}
\value{
An adjacency matrix with an edge added between v1 and v2.
}
\description{
This function adds an edge between two vertices in the given adjacency matrix.
}
\examples{
G_empty <- matrix(0, nrow = 5, ncol = 5)
add_edge(G_empty, 1, 2)

}
