% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA.R
\name{choose_neighbor}
\alias{choose_neighbor}
\title{Randomly choose a neighbor for a given graph.}
\usage{
choose_neighbor(G, d = nrow(G))
}
\arguments{
\item{G}{The graph from which to choose a neighbor.}

\item{d}{The number of nodes in G. Defaults to the number of rows in G.}
}
\value{
The adjacency matrix of the graph with a randomly chosen neighbor.
}
\description{
This function randomly selects a neighbor for a given graph by flipping the state of one randomly chosen edge.
}
\examples{
G <- matrix(1, nrow = 5, ncol = 5) - diag(1, 5)
set.seed(1)
choose_neighbor(G)
}
