% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact.R
\name{exact}
\alias{exact}
\title{Exact algorithm that performs the exhaustive algorithm}
\usage{
exact(sample_df, card_A, d, lambda)
}
\arguments{
\item{sample_df}{sample data}

\item{card_A}{cardinality of the alphabet}

\item{d}{number of vertices}

\item{lambda}{penalty term}
}
\description{
Exact algorithm that performs the exhaustive algorithm
}
