% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact.R
\name{exact_parallel}
\alias{exact_parallel}
\title{Exact algorithm that performs the exhaustive algorithm in parallel}
\usage{
exact_parallel(sample_df, card_A, d, lambda)
}
\arguments{
\item{sample_df}{Sample data.}

\item{card_A}{Cardinality of the alphabet.}

\item{d}{Number of vertices.}

\item{lambda}{Penalty term.}
}
\value{
A list containing the estimated adjacency matrix (G_hat) and the corresponding penalized pseudolikelihood value (logLG_hat).
}
\description{
This function implements an exact algorithm for graph estimation using parallel computing.
}
\examples{
# Example usage
sample_data <- data.frame(V1 = c(0, 1, 0, 1), V2 = c(1, 0, 1, 0), V3 = c(1, 0, 0, 1))
card_A <- 2
d <- 3
lambda <- 0.5
exact_parallel(sample_data, card_A, d, lambda)

}
