% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{generate_and_save_cv_data}
\alias{generate_and_save_cv_data}
\title{Generate and Save Cross-Validation Data Splits}
\usage{
generate_and_save_cv_data(
  original_data,
  n_folds,
  path_to_save,
  file_name,
  seed = 1
)
}
\arguments{
\item{original_data}{The original dataset to be split.}

\item{n_folds}{The number of folds for cross-validation.}

\item{path_to_save}{The path where the generated data splits will be saved.}

\item{file_name}{The name of the file to save the data splits.}

\item{seed}{The random seed for reproducibility (default is 1).}
}
\value{
None (data splits are saved to the specified file).
}
\description{
This function generates cross-validation data splits using the \code{generate_cv_data} function and saves the resulting data to a specified file.
}
