% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{generate_cv_data}
\alias{generate_cv_data}
\title{Generate Cross-Validation Data Splits}
\usage{
generate_cv_data(original_data, n_folds)
}
\arguments{
\item{original_data}{The original dataset to be split.}

\item{n_folds}{The number of folds for cross-validation.}
}
\value{
A tibble containing the generated cross-validation data splits.
The returned tibble has the following columns:
\describe{
  \item{fold}{The fold number.}
  \item{N}{The number of samples in the fold (test data).}
  \item{train}{A list column containing the training data for the fold.}
  \item{test}{A list column containing the test data for the fold.}
}
}
\description{
This function generates data splits for cross-validation.
}
