% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudolikelihood.R
\name{get_and_sum_terms}
\alias{get_and_sum_terms}
\title{Get and sum terms of the pseudolikelihood function}
\usage{
get_and_sum_terms(sample_data, G)
}
\arguments{
\item{sample_data}{Sample data.}

\item{G}{Adjacency matrix of the graph.}
}
\value{
Sum of the terms of the pseudolikelihood function.
}
\description{
This function calculates and sums the terms of the pseudolikelihood function for a given graph represented
by the adjacency matrix G and the sample data. It first obtains the terms using the *get_sum_terms* function
and then sums them using the *sum_terms* function.
}
