% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudolikelihood.R
\name{get_and_sum_terms_with_penalty}
\alias{get_and_sum_terms_with_penalty}
\title{Get and sum terms of the penalized pseudolikelihood function}
\usage{
get_and_sum_terms_with_penalty(sample_data, G, card_A, lambda)
}
\arguments{
\item{sample_data}{Sample data.}

\item{G}{Adjacency matrix of the graph.}

\item{card_A}{Cardinality of the alphabet.}

\item{lambda}{Lambda term in the penalty.}
}
\value{
Sum of the terms of the penalized pseudolikelihood function.
}
\description{
This function calculates and sums the terms of the penalized pseudolikelihood function for a given graph
represented by the adjacency matrix G, the sample data, the cardinality of the alphabet, and the lambda term
in the penalty. It first obtains the terms using the *get_sum_terms* function, then calculates the penalty
using the *penalty* function, and finally subtracts the penalty from the sum of the terms.
}
