% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudolikelihood.R
\name{get_sum_terms}
\alias{get_sum_terms}
\title{Get sum terms of the pseudolikelihood}
\usage{
get_sum_terms(sample_df, G)
}
\arguments{
\item{sample_df}{Data sample.}

\item{G}{Adjacency matrix of the given graph.}
}
\value{
A list of the terms for each vertex in the graph. Each item is a tibble.
}
\description{
This function calculates the sum terms of the pseudolikelihood for each vertex in the graph.
It computes the term for each vertex based on the given sample data and adjacency matrix.
}
