% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_bwd.R
\name{greedy_backwards}
\alias{greedy_backwards}
\title{Implements the greedy backwards stepwise selection algorithm}
\usage{
greedy_backwards(sample_df, lambda, card_A, d, show_all_steps = FALSE)
}
\arguments{
\item{sample_df}{Sample data.}

\item{lambda}{Penalty term.}

\item{card_A}{Cardinality of the alphabet.}

\item{d}{Number of vertices in the graph.}

\item{show_all_steps}{Logical, indicating whether to show all steps of the algorithm.}
}
\value{
A tibble containing the estimated adjacency matrix and its corresponding log-likelihood.
}
\description{
This function implements the greedy backwards stepwise selection algorithm to estimate the graph structure.
It iteratively removes one edge at a time from the initial graph until no improvement in log-likelihood is observed.
}
\examples{
# Example usage
sample_data <- data.frame(V1 = c(1, 2, 3), V2 = c(2, 3, 1))
lambda <- 0.01
card_A <- 3
d <- 2
show_all_steps <- TRUE
greedy_backwards(sample_data, lambda, card_A, d, show_all_steps)

}
