% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_fwd.R
\name{greedy_forward}
\alias{greedy_forward}
\title{Implements the greedy forward stepwise selection algorithm}
\usage{
greedy_forward(sample_df, lambda, card_A, d, show_all_steps = FALSE)
}
\arguments{
\item{sample_df}{Sample data.}

\item{lambda}{Penalty term.}

\item{card_A}{Cardinality of the alphabet.}

\item{d}{Number of vertices in the graph.}

\item{show_all_steps}{Logical indicating whether all steps of the algorithm should be shown.}
}
\value{
A tibble with the estimated adjacency matrix and the corresponding penalized pseudolikelihood value.
}
\description{
This function implements the greedy forward stepwise selection algorithm to estimate the graph.
It iteratively adds one edge at a time to the given adjacency matrix and evaluates the change in the
penalized likelihood. It selects the edge that maximizes the increase in the likelihood and continues
until no more edges can be added without violating the acyclicity constraint of the graph.
}
