% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA.R
\name{identify_diff}
\alias{identify_diff}
\title{Identify differences in the edges of graphs G1 and G2.}
\usage{
identify_diff(G1, G2)
}
\arguments{
\item{G1}{The adjacency matrix of the first graph.}

\item{G2}{The adjacency matrix of the second graph.}
}
\value{
A numeric vector representing the position of the differing edge.
}
\description{
This function identifies the differing edge between two graphs G1 and G2. It assumes that G1 and G2 are neighbors, meaning they differ by only one edge.
}
\examples{
G1 <- matrix(1, nrow = 5, ncol = 5) - diag(1, 5)
set.seed(1)
G2 <- choose_neighbor(G1)
identify_diff(G1, G2)
}
