% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_bwd.R
\name{list_existing_edges}
\alias{list_existing_edges}
\title{List existing edges of an adjacency matrix}
\usage{
list_existing_edges(G)
}
\arguments{
\item{G}{Adjacency matrix of a graph.}
}
\value{
A tibble with a list of all edges in graph G.
}
\description{
This function lists all existing edges in an adjacency matrix.
}
\examples{
# Example usage
G <- matrix(c(0, 1, 1, 0), nrow = 2, ncol = 2)
list_existing_edges(G)

}
