% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_fwd.R
\name{list_possible_edges}
\alias{list_possible_edges}
\title{List possible edges to add to an adjacency matrix}
\usage{
list_possible_edges(G)
}
\arguments{
\item{G}{Adjacency matrix of a graph.}
}
\value{
A tibble with a list of all possible edges to add to graph G.
}
\description{
This function lists all possible edges that can be added to the given adjacency matrix to form a graph.
}
\examples{
G_empty <- matrix(0, nrow = 5, ncol = 5)
list_possible_edges(G_empty)

}
