% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudolikelihood.R
\name{penalty}
\alias{penalty}
\title{Penalty for the likelihood}
\usage{
penalty(lambda, card_A, G)
}
\arguments{
\item{lambda}{Penalty coefficient.}

\item{card_A}{Cardinality of the alphabet.}

\item{G}{Adjacency matrix of the given graph.}
}
\value{
Penalty term for the likelihood.
}
\description{
This function calculates the penalty term for the likelihood. The penalty considered is lambda * |A|^|G|,
where lambda is the penalty coefficient, |A| is the cardinality of the alphabet, and |G| is the number of
edges in the graph represented by the adjacency matrix G.
}
