% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_graph_from_matrix}
\alias{plot_graph_from_matrix}
\title{Plot Graph from Matrix with Fixed Layout}
\usage{
plot_graph_from_matrix(
  G,
  coords,
  node_size = 4,
  title = NULL,
  edge_width = 2,
  edge_color = "darkgray",
  frame_border = TRUE,
  node_font_size = 5,
  label_padding = 0.5,
  alpha_label = 1
)
}
\arguments{
\item{G}{adjacency matrix that represents the graph G}

\item{coords}{matrix of coordinates of the vertices in the plot}

\item{node_size}{node size in the plot}

\item{title}{plot title}

\item{edge_width}{edge width in the plot}

\item{edge_color}{edge color in the plot}

\item{frame_border}{should frame border be shown?}
}
\value{
the graph created from the adjacency matrix
}
\description{
This function plots a graph from an adjacency matrix and node coordinates, using a fixed layout.
}
\examples{
coords <-  matrix(c(1,2, 3,1, 5,2, 4,3, 2,3), ncol = 2, byrow = TRUE)
G <- matrix(1, nrow = 5, ncol = 5) - diag(1, 5)
# plot_graph_from_matrix(G, title = "Fixed random graph", coords = coords)
}
