% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_bwd.R
\name{remove_edge}
\alias{remove_edge}
\title{Remove edge from an adjacency matrix}
\usage{
remove_edge(G, v1, v2)
}
\arguments{
\item{G}{Adjacency matrix.}

\item{v1}{Vertex 1.}

\item{v2}{Vertex 2.}
}
\value{
An adjacency matrix with no edge between v1 and v2.
}
\description{
This function removes the edge between two vertices in an adjacency matrix.
}
\examples{
# Example usage
G_complete <- matrix(1, nrow = 5, ncol = 5) - diag(1, 5)
remove_edge(G_complete, 1, 2)

}
