% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA.R
\name{SimAn}
\alias{SimAn}
\title{Simulated annealing algorithm}
\usage{
SimAn(sample_data, lambda, card_A, G_0, N, C, include_logLG = FALSE)
}
\arguments{
\item{sample_data}{Sample data used for estimation.}

\item{lambda}{Penalty term used in the estimation.}

\item{card_A}{Cardinality of the alphabet.}

\item{G_0}{Initial guess for the adjacency matrix.}

\item{N}{Number of iterations of the simulated annealing algorithm.}

\item{C}{Constant parameter in the simulated annealing algorithm.}

\item{include_logLG}{If TRUE, returns both the estimated adjacency matrix and the corresponding penalized pseudolikelihood value.}
}
\value{
If `include_logLG` is FALSE, returns the estimated adjacency matrix. If TRUE, returns a tibble containing both the estimated adjacency matrix (`G_hat`) and the corresponding penalized pseudolikelihood (`logLG_hat`).
}
\description{
This function implements the simulated annealing algorithm for estimating the adjacency matrix of a graph.
}
