% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_bwd.R
\name{step_down}
\alias{step_down}
\title{Step down of the backwards stepwise greedy algorithm to estimate the graph}
\usage{
step_down(sample_df, G, card_A, lambda)
}
\arguments{
\item{sample_df}{Sample data. The columns should be named as V1, V2, ...}

\item{G}{Adjacency matrix.}

\item{card_A}{Cardinality of the alphabet.}

\item{lambda}{Term in the penalty.}
}
\value{
A tibble containing the adjacency matrix with the removed edge and its corresponding log-likelihood.
}
\description{
This function performs a step down in the backwards stepwise greedy algorithm to estimate the graph.
}
\examples{
# Example usage
sample_data <- data.frame(V1 = c(1, 2, 3), V2 = c(2, 3, 1))
G <- matrix(c(0, 1, 1, 0), nrow = 2, ncol = 2)
card_A <- 3
lambda <- 0.01
step_down(sample_data, G, card_A, lambda)

}
