% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_bwd.R
\name{step_down_parallel}
\alias{step_down_parallel}
\title{Step down of the backwards stepwise greedy algorithm to estimate the graph in parallel}
\usage{
step_down_parallel(sample_df, G, card_A, lambda, clusters = 3)
}
\arguments{
\item{sample_df}{Sample data.}

\item{G}{Adjacency matrix.}

\item{card_A}{Cardinality of the alphabet.}

\item{lambda}{Term in the penalty.}

\item{clusters}{Number of clusters for parallel computation.}
}
\value{
A tibble containing the adjacency matrix with the removed edge and its corresponding log-likelihood.
}
\description{
This function performs a step down in the backwards stepwise greedy algorithm to estimate the graph.
}
