% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_fwd.R
\name{step_up_parallel}
\alias{step_up_parallel}
\title{Step up of the forward stepwise greedy algorithm to estimate the graph in parallel}
\usage{
step_up_parallel(sample_df, G, card_A, lambda, clusters = 3)
}
\arguments{
\item{sample_df}{Sample data.}

\item{G}{Adjacency matrix.}

\item{card_A}{Cardinality of the alphabet.}

\item{lambda}{Term in the penalty.}

\item{clusters}{Number of parallel workers to use.}
}
\value{
A tibble with the updated adjacency matrix and the corresponding log likelihood value.
}
\description{
This function implements the step up of the forward stepwise greedy algorithm to estimate the graph in parallel.
}
