% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudolikelihood.R
\name{sum_terms}
\alias{sum_terms}
\title{sum_terms}
\usage{
sum_terms(terms)
}
\arguments{
\item{terms}{result of *get_sum_terms* function}
}
\value{
a numeric value representing the sum of the terms
}
\description{
sum_terms
}
